move example helper moodules into examples
