+++ enable/examples/demo/enable/basic_draw.py
@@ -11,7 +11,7 @@
 This demonstrates the most basic drawing capabilities using Enable.  A new
 component is created and added to a container.
 """
-from enable.example_support import DemoFrame, demo_main
+from enable.examples.example_support import DemoFrame, demo_main
 from enable.api import Component, Container
