+++ enable/examples/demo/enable/canvas_demo.py
@@ -8,7 +8,7 @@
 #
 # Thanks for using Enthought open source!
 from enable.api import Canvas, Viewport
-from enable.example_support import demo_main, DemoFrame
+from enable.examples.example_support import demo_main, DemoFrame
 from enable.tools.api import ViewportPanTool
