+++ enable/examples/demo/enable/scrolled_demo.py
@@ -14,7 +14,7 @@ from numpy import array
 from traits.api import Enum, Float, Instance, Tuple
 
 from enable.api import Component, Scrolled, Container, Pointer
-from enable.example_support import DemoFrame, demo_main
+from enable.examples.example_support import DemoFrame, demo_main
 
 
 class Circle(Component):
