+++ enable/examples/example_canvas.py
