+++ enable/examples/example_support.py
