+++ kiva/examples/kiva/ellipse.py
@@ -12,7 +12,7 @@ import tempfile
 from scipy import pi
 
 from enable.api import ConstraintsContainer
-from enable.example_support import DemoFrame, demo_main
+from enable.examples.example_support import DemoFrame, demo_main
 from enable.primitives.image import Image
 from kiva.image import GraphicsContext
