+++ kiva/examples/kiva/rect.py
@@ -10,7 +10,7 @@
 import tempfile
 
 from enable.api import ConstraintsContainer
-from enable.example_support import DemoFrame, demo_main
+from enable.examples.example_support import DemoFrame, demo_main
 from enable.primitives.image import Image
 from kiva.image import GraphicsContext
